Refacter pkg/template
