+++ internal/cmd/template/generate.go
@@ -298,6 +298,6 @@ func preTemplateBuild(rawTmpl string) ([]unstructured.Unstructured, error) {
 	inst.SetName("dummy")
 	inst.SetNamespace("dummy")
 
-	builder := template.NewUnstructuredBuilder(rawTmpl, &inst)
+	builder := template.NewRawYAMLBuilder(rawTmpl, &inst)
 	return builder.Build()
 }
