--- pkg/template/vars_test.go
@@ -1,209 +1,0 @@
-package template
-
-import (
-	"reflect"
-	"testing"
-
-	cosmov1alpha1 "github.com/cosmo-workspace/cosmo/api/core/v1alpha1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-func TestUnstructuredBuilder_ReplaceDefaultVars(t *testing.T) {
-	type fields struct {
-		rawYaml string
-		inst    *cosmov1alpha1.Instance
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		want   *UnstructuredBuilder
-	}{
-		{
-			name: "OK",
-			fields: fields{
-				rawYaml: "{{INSTANCE}}-{{NAMESPACE}}-{{TEMPLATE}}",
-				inst: &cosmov1alpha1.Instance{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cs1",
-						Namespace: "cosmo-user-tom",
-					},
-					Spec: cosmov1alpha1.InstanceSpec{
-						Template: cosmov1alpha1.TemplateRef{
-							Name: "code-server",
-						},
-						Override: cosmov1alpha1.OverrideSpec{},
-						Vars:     map[string]string{"{{TEST}}": "OK"},
-					},
-				},
-			},
-			want: &UnstructuredBuilder{
-				rawYaml: "cs1-cosmo-user-tom-code-server",
-				inst: &cosmov1alpha1.Instance{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cs1",
-						Namespace: "cosmo-user-tom",
-					},
-					Spec: cosmov1alpha1.InstanceSpec{
-						Template: cosmov1alpha1.TemplateRef{
-							Name: "code-server",
-						},
-						Override: cosmov1alpha1.OverrideSpec{},
-						Vars:     map[string]string{"{{TEST}}": "OK"},
-					},
-				},
-			},
-		},
-		{
-			name: "without brackets",
-			fields: fields{
-				rawYaml: "{{INSTANCE}}-{{NAMESPACE}}-{{TEMPLATE}}",
-				inst: &cosmov1alpha1.Instance{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cs1",
-						Namespace: "cosmo-user-tom",
-					},
-					Spec: cosmov1alpha1.InstanceSpec{
-						Template: cosmov1alpha1.TemplateRef{
-							Name: "code-server",
-						},
-						Override: cosmov1alpha1.OverrideSpec{},
-						Vars:     map[string]string{"TEST": "OK"},
-					},
-				},
-			},
-			want: &UnstructuredBuilder{
-				rawYaml: "cs1-cosmo-user-tom-code-server",
-				inst: &cosmov1alpha1.Instance{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cs1",
-						Namespace: "cosmo-user-tom",
-					},
-					Spec: cosmov1alpha1.InstanceSpec{
-						Template: cosmov1alpha1.TemplateRef{
-							Name: "code-server",
-						},
-						Override: cosmov1alpha1.OverrideSpec{},
-						Vars:     map[string]string{"TEST": "OK"},
-					},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tr := &UnstructuredBuilder{
-				rawYaml: tt.fields.rawYaml,
-				inst:    tt.fields.inst,
-			}
-			if got := tr.ReplaceDefaultVars(); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("UnstructuredBuilder.ReplaceDefaultVars() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestUnstructuredBuilder_ReplaceCustomVars(t *testing.T) {
-	type fields struct {
-		rawYaml string
-		inst    *cosmov1alpha1.Instance
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		want   *UnstructuredBuilder
-	}{
-		{
-			name: "OK",
-			fields: fields{
-				rawYaml: "{{INSTANCE}}-{{NAMESPACE}}-{{TEMPLATE}}-{{TEST}}",
-				inst: &cosmov1alpha1.Instance{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cs1",
-						Namespace: "cosmo-user-tom",
-					},
-					Spec: cosmov1alpha1.InstanceSpec{
-						Template: cosmov1alpha1.TemplateRef{
-							Name: "code-server",
-						},
-						Override: cosmov1alpha1.OverrideSpec{},
-						Vars:     map[string]string{"{{TEST}}": "OK"},
-					},
-				},
-			},
-			want: &UnstructuredBuilder{
-				rawYaml: "{{INSTANCE}}-{{NAMESPACE}}-{{TEMPLATE}}-OK",
-				inst: &cosmov1alpha1.Instance{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cs1",
-						Namespace: "cosmo-user-tom",
-					},
-					Spec: cosmov1alpha1.InstanceSpec{
-						Template: cosmov1alpha1.TemplateRef{
-							Name: "code-server",
-						},
-						Override: cosmov1alpha1.OverrideSpec{},
-						Vars:     map[string]string{"{{TEST}}": "OK"},
-					},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tr := &UnstructuredBuilder{
-				rawYaml: tt.fields.rawYaml,
-				inst:    tt.fields.inst,
-			}
-			if got := tr.ReplaceCustomVars(); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("UnstructuredBuilder.ReplaceCustomVars() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestValidCustomVars(t *testing.T) {
-	type args struct {
-		varString string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "OK",
-			args: args{
-				varString: "{{INSTACE}}",
-			},
-			wantErr: false,
-		},
-		{
-			name: "Invalid",
-			args: args{
-				varString: "INSTACE",
-			},
-			wantErr: true,
-		},
-		{
-			name: "Invalid sufix",
-			args: args{
-				varString: "{{INSTACE)",
-			},
-			wantErr: true,
-		},
-		{
-			name: "Invalid prefix",
-			args: args{
-				varString: "$(INSTACE}}",
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := ValidCustomVars(tt.args.varString); (err != nil) != tt.wantErr {
-				t.Errorf("ValidCustomVars() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
